Fixed alarm sound not playing

+++ src/timer.js
@@ -3,7 +3,7 @@ import { Howl } from "howler";
 import { alarmSound } from "./main";
 import { data } from "autoprefixer";
 
-export let worker = new Worker(new URL("worker.js", import.meta.url));
+export let worker = new Worker(new URL("worker.js", import.meta.url ));
 
 export function timer() {
   const pad = '0'
@@ -20,16 +20,42 @@ export function timer() {
     // and create a new one called worker because the other 'worker' is destroyed
     // you can reaasign it                                                          
   }
-
   timerWorker(minutes, seconds);
+}
+
+export function timerWorker(minutes , seconds) {
+  let strmin = minutes.textContent;
+  let strsec = seconds.textContent;
+  if (window.Worker) { // checks if web workers are supported in the browser
+
+    worker.onmessage = function(event) { 
+      // this is the data that we received from the worker.js sent
+      // the onmessage is a event listener
+      // this is the endpoit of the worker function 
+      // all that has modified or changed? will be sent here
+      // like if the min and sec were 0 then do that
+      // and by the end stop the worker
+        const { min,sec,intervalId } = event.data;
+        console.log(min,sec)
+        
+        minutes.textContent = min
+        seconds.textContent = sec
+        if(minutes.textContent === '0' && seconds.textContent === '00'){
+            alarmSound.play()
+            worker.terminate();
+            return;
+        }
+        return intervalId;
+    }; 
   
-  // let intervalId = setInterval(() => {
-  if(minutes.textContent === '0' && seconds.textContent === '00') {
-    alarmSound.play()
-    worker.terminate();
-    console.log("oloy")
-    return 
-  }  
+    worker.postMessage({strmin , strsec});  // this is the data that we will send to the worker.js
+  } else {
+    console.log("Web Workers are not supported in this browser.");
+  }
+
+}
+// let intervalId = setInterval(() => {
+  
   //   if(minutes.textContent === '0' && seconds.textContent === '00') {
   //     alarmSound.play()
   //     clearInterval(intervalId)
@@ -48,28 +74,3 @@ export function timer() {
   //     seconds.textContent = pad + numSecs;
   //   }
   // }, 1000);
-}
-
-export function timerWorker(minutes , seconds) {
-  let strmin = minutes.textContent;
-  let strsec = seconds.textContent;
-  if (window.Worker) { // checks if web workers are supported in the browser
-
-    worker.onmessage = function(event) { 
-      // this is the data that we received from the worker.js sent
-      // the onmessage is a event listener
-        const { min,sec,intervalId } = event.data;
-        console.log(min,sec)
-
-        minutes.textContent = min
-        seconds.textContent = sec
-
-        return intervalId;
-    }; 
-  
-    worker.postMessage({strmin , strsec});  // this is the data that we will send to the worker.js
-  } else {
-    console.log("Web Workers are not supported in this browser.");
-  }
-
-}
